Fold tiny plot/core/defaults.ts into axis-utils.ts
AXIS_DEFAULTS (the only export) now lives alongside the axis utilities that use
it; ScatterPlot/Histogram import it from there. Keeps the file tree leaner with
no public-API change (defaults was deep-imported, not barrel-exported).
