+++ src/lib/plot/histogram/Histogram.svelte
@@ -18,9 +18,8 @@
     ReferenceLine,
   } from '$lib/plot'
   import type { AxisChangeState } from '$lib/plot/core/axis-utils'
-  import { create_axis_change_handler } from '$lib/plot/core/axis-utils'
+  import { AXIS_DEFAULTS, create_axis_change_handler } from '$lib/plot/core/axis-utils'
   import { extract_series_color, prepare_legend_data } from '$lib/plot/core/data-transform'
-  import { AXIS_DEFAULTS } from '$lib/plot/core/defaults'
   import {
     create_dimension_tracker,
     create_hover_lock,
